+++ client/components/Main/Mainsection/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useRecoilValue } from 'recoil';
 import Link from 'next/link';
 import ArticleCard from '../Articlecard';
-import newsfeedState from '../../../atom/newsfeedState';
+import { newsfeedState } from '../../../atom';
 import { ArticlesSection, FakeButton, NewArticleSection, Placeholder, Wrapper } from './index.style';
 
 export default function MainSection() {
