+++ server/src/main.ts
@@ -22,7 +22,7 @@ async function bootstrap() {
       errorHttpStatusCode: HttpStatus.UNPROCESSABLE_ENTITY,
     }),
   );
-  
+
   await app.listen(4000);
 }
 bootstrap();
